chore(test): refactor create mock entities context
