+++ packages/ui/src/components/Visualization/Custom/hooks/disable-step.hook.test.tsx
@@ -6,6 +6,7 @@ import { CamelRouteResource } from '../../../../models/camel/camel-route-resourc
 import { IVisualizationNode } from '../../../../models/visualization/base-visual-entity';
 import { createVisualizationNode } from '../../../../models/visualization/visualization-node';
 import { EntitiesContext, EntitiesContextResult } from '../../../../providers/entities.provider';
+import { createMockEntitiesContext } from '../../../../stubs/create-mock-entities-context';
 import { useDisableStep } from './disable-step.hook';
 
 vi.mock('@kaoto/forms', () => ({
@@ -18,15 +19,7 @@ describe('useDisableStep', () => {
   let mockEntitiesContext: EntitiesContextResult;
 
   beforeAll(async () => {
-    await camelResource.initialize();
-    mockEntitiesContext = {
-      camelResource,
-      entities: camelResource.getEntities(),
-      visualEntities: camelResource.getVisualEntities(),
-      currentSchemaType: camelResource.getType(),
-      updateSourceCodeFromEntities: vi.fn(),
-      updateEntitiesFromCamelResource: vi.fn(),
-    };
+    mockEntitiesContext = await createMockEntitiesContext(camelResource);
   });
 
   beforeEach(() => {
